Create a unified object logger (#94)
The object logger will be used for various debug logging use-cases, including replacing training history and replay buffer logging.

The logger buffers and pickle's logged entries and appends them to a file. The reader is an iterable, hiding the fact that the entries were written in batches.

ObjectLogManager provides the actual unified log experience. New types of entries can be logged by simply calling log with the log_filename (which also serves as a label) and the new entries themselves.

Standard users will not instantiate ObjectLogger directly.

The BridgeBuilderModel __init__ will be modified to take an ObjectLogManager instance as an argument to store into self.