Log normalizer (#102)
Some objects are expensive to log repeatedly. The LoggerAndNormalizer provides functionality to log these expensive objects exactly once and provide a per-execution unique id for each object to be used in other, more frequent log entries.

That is, if the logged files were a database schema, the frequently written 'fact' table and the 'dimension' table with the expensive entity would be considered to be normalized.

One expected use case is to log state exactly once per state encountered and use the id when logging other aspects of the training history, including q values which may be computed many times per state.

This PR continues the migration to the new logging system.