Log normalize states (#107)
Utilize LoggerAndNormalizer for states in builder_trainer to reduce the logging overhead from the states and next_states tensors. The first beneficiary is the TrainingBatchLogEntry.

I used the following function for measuring speed-ups. After changing the env to width=6, max_episode_length=100, and batch-size=500, the logging cost reduced by 6x with and without this PR's changes.