Fix numpy to Tensor conversion to maintain types. (#121)
Converting the numpy state array to floats also causes the same state to appear to have two different representations/ids when hashed and normalized because the int and floats version of the state appear to be different. Fixing the types usage eliminates this, meaning there is one fewer "observed" state according to the LoggerAndNormalizer in the unit tests. That is why several state ids have been updated to be -1 of their previous value.