Action inversion checker (#141)
The action inversion checker identifies when a policy used to identify a preferred action for a state and then stopped. It will be part of our tool set in debugging why/when the policies starts diverging badly after converging.

This PR submits a standalone utility tied to current bridge building problems.

Future PRs will:

1. Integrate the action inversion checker into the builder_trainer
2. Log the ActionInversionReports via object logger
3. Create a simple visualization or log browse utility to debug.

This PR also creates a debug directory into which we may move other tools.