Q interface (#174)
This PR is the first in a line of 3 to implement Tabular Q learning as another Q function option.

Here, we create a unified interface for a QManager, which can be implemented in multiple ways. The QManager encapsulates the q and target networks together, strengthening the promise that they will carry the same architecture and be compatible.