bar. (#182)
Adds a green bar to the top of the screen to indicate when data is loading and a red one when there's a load error.

Sometimes Sibyl takes time to load all the data for a request and this makes it clear to the user when the plots are updated per request. We use absolute positioning to avoid shifting the page up and down.

The load error indicator requires user investigation in the server log.