done. (#180)
Instead of replacing the object_logging dir with the most recent run each time, we know store all experiment runs in a single base directory. A new subdir is created per experiment run, named after the datetime and an optional experiment name.

The next PR addresses the required changes for Sibyl. As is, this change does not break Sibyl. It just changes which dir is provided as the --log_dir flag value when using Sibyl.