Support refreshable/pasetable URLs for Training History Viewer parameters  (#184)
Whenever the plots are updated, this PR will store all the parameters (eg start_batch_idx, the state body filter function, etc) into the page URL. Therefore, if the user refreshes the page or copy-pastes the URL to a different tab, the exact same view will be loaded, preserving all the filter parameters. When we have sibyl hosted on a remote server, this means the training history viewer urls can also be shared between people to show the same view.

Main changes:
* Redefine get_string_or_none and get_int_or_none to return default values.
* Change the page to accept GET params and set those in the HTML using the template
* Update the URL whenever new parameters are changed and submitted/applied.

I have not updated the action inversion page at this time. That page is less used and unclear how long it will be useful. I'm opting for a 'update it when it feels necessary' approach.