Direct load training history database (#198)
Reduces Sibyl load times from 11-12s to 600-700ms for the test dataset.

Building a TrainingHistoryDatabase and ActionInversionDatabase from logs takes a significant fraction of the time in the process of loading data into SIbyl. Once built, the database can be pickled and reloaded ~15x faster. Here we add an intermediate step during cache miss to first checked if a database has already been loaded saved from logs. If so, we use the faster path to re-load it again directly. If not, the database must be built from logs and will be slower. That database will also be pickled for future Sibyl runs. These built and saved databases persist across Sibyl runs.

A following CL will add a background thread to build the databases from logs.

A future CL could consider removing the in-memory portion of the ObjectLogCache if we're worried about memory usage. Currently we observe 600-700ms load times from disk reduced to ~50-60ms. The rendering of charts on the browser now dominates the e2e load time for the user and so this 10x savings between loading files and using in-memory cache may not be worth the memory management/cache eviction overhead to ensure the cache doesn't grow too large.