Fix data loader (#216)
Upgrade lightning
Change to use epochs as 1 epoch per iteration instead of many batches in a single epoch so that the data loader can be reset each epoch and pick up changes in the replay buffer. python iterators, like in the data loader, are immutable and don't pick changes in the replay buffer so we have never used the replay buffer's new memories so far.