Add a panel to sibyl to show which batches contained at least one sample of a (state, action) pair (#222)
Use the training batch log entry to plot for each state row in sibyl, a scatter plot with the x axis as the batch idx and y axis as the action if the action was sampled for that batch idx.

Includes a light refactor to make plot type setting when the metrics are defined.

Adjust CSS to keep a single row per state. in the future, this could be multi-row or we can remove one of the Q or Q-Target plots, perhaps? (if this feels hard to read).

Note: You may need to clear your sibyl_tmp